refactor: move wordpress-seo code to separate class
