+++ includes/plugins/class-wordpress-seo.php
@@ -1,0 +1,52 @@
+<?php
+/**
+ * Yoast (wordpress-seo) integration class.
+ *
+ * @package Newspack
+ */
+
+namespace Newspack;
+
+defined( 'ABSPATH' ) || exit;
+
+/**
+ * Main class.
+ */
+class WordPress_SEO {
+	/**
+	 * Initialize hooks and filters.
+	 */
+	public static function init() {
+		add_filter( 'wpseo_enhanced_slack_data', [ __CLASS__, 'use_cap_for_slack_preview' ] );
+		add_filter( 'wpseo_opengraph_url', [ __CLASS__, 'http_ogurls' ] );
+	}
+
+	/**
+	 * Use the Co-Author in Slack preview metadata instead of the regular post author if needed.
+	 *
+	 * @param array $slack_data Array of data which will be output in twitter:data tags.
+	 * @return array Modified $slack_data
+	 */
+	public static function use_cap_for_slack_preview( $slack_data ) {
+		if ( function_exists( 'coauthors' ) && is_single() && isset( $slack_data[ __( 'Written by', 'wordpress-seo' ) ] ) ) {
+			$slack_data[ __( 'Written by', 'wordpress-seo' ) ] = coauthors( null, null, null, null, false );
+		}
+
+		return $slack_data;
+	}
+
+	/**
+	 * On Atomic infrastructure, URLs are `http` for Facebook requests.
+	 * This forces the `og:url` to `http` for consistency, to prevent 301 redirect loop issues.
+	 *
+	 * @param string $og_url The opengraph URL.
+	 * @return string modified $og_url
+	 */
+	public static function http_ogurls( $og_url ) {
+		if ( defined( 'ATOMIC_SITE_ID' ) && ATOMIC_SITE_ID ) {
+			$og_url = str_replace( 'https:', 'http:', $og_url );
+		}
+		return $og_url;
+	}
+}
+WordPress_SEO::init();
